crawler/requirements.txt
battlerite-client==0.1.2
python-dotenv>=0.7
